Add benchmark info to the footer message of say blocks
This change is modeled after Rails migrations (the role model for this
whole gem). There, for example, we see:

```
==  CreateProducts: migrating =================================================
-- create_table(:products)
   -> 0.0028s
==  CreateProducts: migrated (0.0028s) =============================
```

So, likewise for `say_with_block` (basically, any time we have info on
how long it took for something to run), we'll now include benchmark
info. e.g. instead of just:

```
= Done =========================================================================
```

We'll now display, e.g.:

```
= Done (0.0000s) ===============================================================
```