Refactor Say's Public API methods
Remove `build_` prefix
- `Say.build_banner` -> `Say.say_banner`
- `Say.build_message` -> `Say.say_message`

This is consistent with how we don't have `Say.build_say` and with how
all the other methods are prefixed with `say_`.

Prefixing with `say_` defines the public API of this gem. And this is
important because of the ability to use `include Say` -- we have to
namespace instance methods with the gem name to prevent collisions.

Rename `Say.say_item` -> `Say.say_result`
- This is a better description of what's happening in this method, 
  especially when paired with the idea of a "type" (a result type).
  i.e. the point of this gem is to output processing results (and to
  demarcate logical processing blocks into visible output blocks).

Also, rename the `message` method args and local vars to `text` for
clarity.