Rename test/test_say.rb -> test/say_test.rb
This is prep for expanding the Say namespace. I'm not sure what name to
use for testing e.g. Say::Progress if the convention is Test<Class> (the
bundler default/convention) vs <Class>Test (the Rails
default/convention). Would it be `Say::TestProgress` or
`TestSay::Progress`? Both are confusing, IMO.
- I prefer Say::ProgressTest because, at least this way, the namespace
  and class name are more clear vs the "Test" type identifier.