Add `amazing_print` gem as a development dependency
I'll be using this for prettier output when running test methods (not
unit testing, but the test_ methods defined on objects for easier,
predictable on-the-console testing or new features, etc. This is a bit
like acceptance testing or feature testing, I guess, but it's a pattern
I learned from a SmallTalk developer and I like it.).