Remove arg `silent` from Say.write method
"Silent Say" is a carry-over from the project this gem was extracted
from. It's not really relevant, however, as I had temporarily
implemented it here (as a control variable on Say.write). The idea of a
"Silent Say" is supposed to be one in which "Say" output is not
silenced, but, rather, Logger output is silenced. This is outside the
scope of this gem (at least for now... though it could be perhaps be
supported later using callbacks or something).