Change the default interpolation template string to "{}"
This old default was both arbitrary and yet indicated a preference for
use in left-justified banners / :hr banners.

This was bad because it acted as conceptual coupling... which I've found
to be difficult to keep straight. i.e. which objects are expecting which
other objects to act in which way?

So, with the new default interpolation template string ("{}") no
flourishes are added on top of the default interpolation sentinel, in an
attempt to be neutral or to "know nothing" about "preferred" use cases.

Also, fix the tests that ensured the various interrogation results of
the default interpolation template string. They'll now be duplicates of
the tests for the "empty" interpolation template string tests, but I
like this duplication because it keeps things explicit and serves as
obvious placeholders for testing changes if the default interpolation
template string does ever change again in the future.