Complete rewrite of Say::InterpolationTemplate
This is a break in the normal release cadence... all code changes
between v0.4.0 and this commit are now considered an internal-only
"release". I've left the CHANGELOG entries for these but marked them as
an Unreleased Version. This is due to this commit containing a major
change to how banners, justification, and interpolation templates are
handled. I couldn't see my way through iterative change on this as it's
just a personal project with no real users at this point.

Rewrite Say::InterpolationTemplate
- Say::InterpolationTemplate is now defined through piecemeal
  attributes, instead of by passing in a representative interpolation
  template String. This allows for very obvious and fine-grained control
  of which parts of the template are which.
- Add the ability to specify left/right bookends on
  Say::InterpolationTemplate. These allow for specifying text that
  should always appear on the left/right end of the interpolated text,
  regardless of length restrictions.

Remove the Say::LJBanner, Say::CJBanner, and Say::RJBanner classes.
- Instead, we now have Say::JustificationBehaviors and implementing
  classes: Say::LeftJustifier, Say::CenterJustifier, and
  Say::RightJustifier.
- These new classes are internally instantiated when calling
  Say::InterpolationTemplate#left_justify,
  Say::InterpolationTemplate#center_justify, and
  Say::InterpolationTemplate#right_justify, respectively.

Update README to show advanced usage using the newly rewritten
Say::InterpolationTemplate class.